Setup Toolbar
Note that when the android:fitsSystemWindows attribute is set to true
for a view, the view would be laid out as if the StatusBar and the
ActionBar were present i.e. the UI on top gets padding enough to not be
obscured by the navigation bar. Without this attribute, there is not
enough padding factored into consideration for the ToolBar:

We want our main content view to have the navigation bar and hence
android:fitsSystemWindows is set to true for the Toolbar.
To use the Toolbar as an ActionBar, you need to disable the default
ActionBar. This can be done by setting the app theme in styles.xml file.

<resources>
<!-- Base application theme. -->
<style name="AppTheme" parent="Theme.AppCompat.Light.NoActionBar">
<item name="colorPrimary">#673AB7</item>
<item name="colorPrimaryDark">#512DA8</item>
<item name="colorAccent">#FF4081</item>
</style>
</resources>
Note: If you forget to disable the ActionBar in styles.xml, you are
likely to see a java.lang.IllegalStateException with an error message
that reads This Activity already has an action bar supplied by the
window decor. Do not request Window.FEATURE_ACTION_BAR and set
windowActionBar to false in your theme to use a Toolbar instead. If you
see this message, you need to make sure to follow the previous steps.