Add Navigation Header
The NavigationView also accepts a custom attribute that can reference a
layout that provides a header of our layout. For instance, you can
create a layout/nav_header.xml similar to the following:
You would then reference this in the layout res/layout/activity_main.xml
in the NavigationView with the app:headerLayout custom attribute:
This app:headerLayout inflates the specified layout into the header
automatically.