Animate the Hamburger icon
I also fixed the second fragment. It appears I accidentally created
callbacks in it.
In order for the hamburger icon to animate to indicate the drawer is
being opened and closed, we need to use the ActionBarDrawerToggle class.
We need to tie the DrawerLayout and Toolbar together:
Next, we need to make sure we synchronize the state whenever the screen
is restored or there is a configuration change (i.e screen rotation):
We also need to change the onOptionsItemSelected() method and allow the
ActionBarToggle to handle the events.
The ActionBarToggle will perform the same function done previously but
adds a bit more checks and allows mouse clicks on the icon to open and
close the drawer. See the source code for more context.
One thing to note is that the ActionBarDrawerToggle renders a custom
DrawerArrowDrawable for you for the hamburger icon.